+++ src/configurator/__tests__/index.js
@@ -18,7 +18,7 @@ test("should return a default configuration", (t) => {
     nojekyll: true,
     devHost: "0.0.0.0",
     devPort: 3000,
-    appcache : false,
+    offline : false,
     verbose: false,
     open: true,
     dev: false,
@@ -39,6 +39,8 @@ test("should return a default configuration", (t) => {
       path: "/",
       href: "http://0.0.0.0:3000/",
     },
+    // deprecated
+    appcache: undefined,
   }
 
   t.deepEqual(
@@ -98,111 +100,4 @@ test("should adjust 'NODE_ENV' when '--production' is used", (t) => {
   process.env.NODE_ENV = undefined
   configurator({ homepage: "http://a.b/" }, [ "--production" ])
   t.is(process.env.NODE_ENV, "production")
-})
-
-test("should accept string for 'asset' option", (t) => {
-  const config = configurator(
-    {
-      phenomic: {
-        "assets": "AsSeT",
-      },
-    },
-    []
-  )
-  t.deepEqual(
-    config.assets,
-    {
-      path: join(process.cwd(), config.source, "AsSeT"),
-      route:"AsSeT",
-    }
-  )
-})
-
-test("should accept true for 'asset' option", (t) => {
-  const config = configurator(
-    {
-      phenomic: {
-        "assets": true,
-      },
-    },
-    []
-  )
-  t.deepEqual(
-    config.assets,
-    {
-      path: join(process.cwd(), config.source, "assets"),
-      route:"assets",
-    }
-  )
-})
-
-test("should not accept false for 'asset' option", (t) => {
-  const config = configurator({ phenomic: { "assets": false } }, [])
-  t.is(config.assets, false)
-})
-
-test("should not accept null for 'asset' option", (t) => {
-  const config = configurator({ phenomic: { "assets": null } }, [])
-  t.is(config.assets, null)
-})
-
-test("should accept object for 'asset' option", (t) => {
-  t.throws(
-    () => {
-      configurator({ phenomic: { assets: { } } }, [])
-    },
-    (error) => error.message.includes(
-      "You provided an object for 'assets' option."
-    )
-  )
-})
-
-test("should default to false for 'appcache' option", (t) => {
-  t.is(
-    configurator({}, []).appcache,
-    false
-  )
-})
-
-test("should accept string for 'appcache' option", (t) => {
-  t.deepEqual(
-    configurator({ phenomic: { appcache: "foo" } }, []).appcache,
-    [ "foo" ]
-  )
-})
-
-test("should accept array for 'appcache' option", (t) => {
-  t.deepEqual(
-    configurator({ phenomic: { appcache: [ "foo" ] } }, []).appcache,
-    [ "foo" ]
-  )
-})
-
-test("should return default config when 'appcache' is true", (t) => {
-  t.deepEqual(
-    configurator({ phenomic: { appcache: true } }, []).appcache,
-    [ "**/*.*", "!**/*.html", "index.html" ]
-  )
-})
-
-test("should accept falsy for 'appcache' option", (t) => {
-  t.is(
-    configurator({ phenomic: { appcache: false } }, []).appcache,
-    false
-  )
-  t.is(
-    configurator({ phenomic: { appcache: null } }, []).appcache,
-    false
-  )
-})
-
-test("should not accept object for 'appcache' option", (t) => {
-  t.throws(
-    () => {
-      configurator({ phenomic: { appcache: { foo: "bar" } } }, [])
-    },
-    (error) => error.message.includes(
-      "You provided an 'object' for 'appcache' option."
-    )
-  )
 })