+++ src/configurator/validators.js
@@ -1,4 +1,4 @@
-export appcache from "./validators/appcache.js"
+export offline from "./validators/offline.js"
 export assets from "./validators/assets.js"
 export baseUrl from "./validators/baseUrl.js"
 export production from "./validators/production.js"